+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Major: Add boss kill count notifications. (#50)
+
 ## 1.0.3
 
 - Bugfix: Ensure send image configs are adhered to. (#51)